add funct to wait for influx service to start
